+++ MapDriveInator/MapDriveInator.mm
@@ -192,7 +192,8 @@
       To translate the description you have to define a translation for the key 'addons.${name}.description'.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="AddOn to coordinate the disk&apos;s folder structure with the nodes structure in the mind map (back and forth)" ID="ID_659221198"/>
 </node>
 <node TEXT="changes" POSITION="top_or_left" ID="ID_1078107435">
@@ -207,7 +208,8 @@
       Change log of this add-on: append one node for each noteworthy version and put the details for each version into a child node.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="v0.0.3" FOLDED="true" ID="ID_1486725916">
 <node TEXT="old script version packed as an Add-on" ID="ID_1092003668"/>
 </node>
@@ -224,6 +226,7 @@
 </node>
 <node TEXT="v0.0.6" ID="ID_463111565">
 <node TEXT="MapDriveInator: nodes with style file_folder, loose this style after the folder is created in drive and they get the link to it" ID="ID_1694158485"/>
+<node TEXT="Bug: [&quot;Allow nodes to have links&quot;](https://github.com/EdoFro/Freeplane_Map-Drive-Inator/issues/15) solved" ID="ID_1645231008"/>
 </node>
 </node>
 <node TEXT="license" FOLDED="true" POSITION="top_or_left" ID="ID_1028448710">
@@ -244,7 +247,8 @@
       The License text has to be entered as a child of the <i>'license'</i>&nbsp;node, either as plain text or as HTML.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="MIT License&#xa;&#xa;Copyright (c) 2021 Eduardo Frohlich.&#xa;&#xa;Permission is hereby granted, free of charge, to any person obtaining a copy&#xa;of this software and associated documentation files (the &quot;Software&quot;), to deal&#xa;in the Software without restriction, including without limitation the rights&#xa;to use, copy, modify, merge, publish, distribute, sublicense, and/or sell&#xa;copies of the Software, and to permit persons to whom the Software is&#xa;furnished to do so, subject to the following conditions:&#xa;&#xa;The above copyright notice and this permission notice shall be included in all&#xa;copies or substantial portions of the Software.&#xa;&#xa;THE SOFTWARE IS PROVIDED &quot;AS IS&quot;, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR&#xa;IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,&#xa;FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE&#xa;AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER&#xa;LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,&#xa;OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE&#xa;SOFTWARE." ID="ID_1379820487" VSHIFT_QUANTITY="-0.75 pt"/>
 </node>
 <node TEXT="preferences.xml" POSITION="top_or_left" ID="ID_720156944">
@@ -355,7 +359,8 @@
       </tr>
     </table>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 </node>
 <node TEXT="default.properties" POSITION="top_or_left" ID="ID_536060528">
 <edge COLOR="#00ffff"/>
@@ -377,7 +382,8 @@
       </li>
     </ul>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <attribute_layout NAME_WIDTH="80.25 pt" VALUE_WIDTH="122.25 pt"/>
 </node>
 <node TEXT="translations" POSITION="top_or_left" ID="ID_137723580">
@@ -412,7 +418,8 @@
       </li>
     </ul>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="en" ID="ID_1122736476">
 <attribute_layout NAME_WIDTH="196.49999 pt" VALUE_WIDTH="140.25 pt"/>
 <attribute NAME="addons.${name}" VALUE="MapDriveInator"/>
@@ -445,7 +452,8 @@
       List of files and/or directories to remove on uninstall
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <attribute_layout NAME_WIDTH="40.5 pt" VALUE_WIDTH="287.24999 pt"/>
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}.script.xml"/>
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}/scripts/checkMissingFileNodes.groovy"/>
@@ -583,7 +591,8 @@
       &nbsp;&nbsp;- In any case set execute_scripts_without_asking to true unless you want to annoy users.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="checkMissingFileNodes.groovy" ID="ID_953928223">
 <attribute_layout NAME_WIDTH="183.74999 pt" VALUE_WIDTH="164.25 pt"/>
 <attribute NAME="menuTitleKey" VALUE="addons.${name}.checkMissingFileNodes"/>
@@ -801,7 +810,8 @@
       &nbsp;- The files will be processed in the sequence as seen in the map.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="MapDriveInator.jar" ID="ID_1212321759"/>
 </node>
 <node TEXT="zips" POSITION="bottom_or_right" ID="ID_26433714">
@@ -840,7 +850,8 @@
       &nbsp;- The files will be processed in the sequence as seen in the map.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="doc" ID="ID_1305821302"/>
 <node TEXT="icons" ID="ID_615436065"/>
 <node TEXT="templates" ID="ID_903379418"/>
@@ -877,7 +888,8 @@
       Images can be added automatically by releaseAddOn.groovy or must be uploaded into the map via the script <i>Tools-&gt;Scripts-&gt;Insert Binary</i>&nbsp;since they have to be (base64) encoded as simple strings.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 </node>
 <node TEXT="actions" POSITION="bottom_or_right" ID="ID_1052924007"><richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
 <html>
@@ -887,7 +899,8 @@
   <body>
     Direct links to menu commands
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="Build add-on" ID="ID_14185460" LINK="menuitem:_addons.devtools.checkAddOn_on_single_node"/>
 <node TEXT="Package add-on for publication" ID="ID_1767675080" LINK="menuitem:_addons.devtools.releaseAddOn_on_single_node"/>
 <node TEXT="Export Translations" ID="ID_1894374378" LINK="menuitem:_addons.devtools.exportTranslations_on_single_node"/>